Change to new save format
Managing my collection of shader JSON files has been difficult because
the files give no indication of what their contents look like. The new
save format stores the JSON in the pixel data of an image right below a
picture of the shader. I considered saving the JSON in a PNG tEXt chunk,
but concluded that using invisible metadata would be confusing.

The images contain a watermark to point the user back to the application
where they can be edited.